+++ code/core/src/main/java/org/betonquest/betonquest/database/AsyncSaver.java
@@ -9,7 +9,7 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 /**
  * Saves the data to the database asynchronously.
  */
-@SuppressWarnings({"PMD.DoNotUseThreads", "PMD.AvoidSynchronizedStatement"})
+@SuppressWarnings("PMD.AvoidSynchronizedStatement")
 @SuppressFBWarnings("IS2_INCONSISTENT_SYNC")
 public class AsyncSaver implements Runnable, Saver {
 
@@ -29,11 +29,6 @@ public class AsyncSaver implements Runnable, Saver {
     private final Queue<Record> queue;
 
     /**
-     * The amount of time until the AsyncSaver tries to reconnect if there was a connection loss.
-     */
-    private final long reconnectInterval;
-
-    /**
      * Whether the saver is currently running or not.
      */
     private boolean running;
@@ -41,13 +36,11 @@ public class AsyncSaver implements Runnable, Saver {
     /**
      * Creates a new database saver thread.
      *
-     * @param log               the logger that will be used for logging
-     * @param reconnectInterval the interval for trying reconnecting to the database
-     * @param connector         the connector for database access
+     * @param log       the logger that will be used for logging
+     * @param connector the connector for database access
      */
-    public AsyncSaver(final BetonQuestLogger log, final long reconnectInterval, final Connector connector) {
+    public AsyncSaver(final BetonQuestLogger log, final Connector connector) {
         this.log = log;
-        this.reconnectInterval = reconnectInterval;
         this.con = connector;
         this.queue = new ConcurrentLinkedQueue<>();
         this.running = true;
@@ -55,9 +48,7 @@ public class AsyncSaver implements Runnable, Saver {
 
     @Override
     @SuppressFBWarnings("UW_UNCOND_WAIT")
-    @SuppressWarnings("PMD.CognitiveComplexity")
     public void run() {
-        boolean active = false;
         while (true) {
             while (queue.isEmpty()) {
                 if (!running) {
@@ -65,30 +56,16 @@ public class AsyncSaver implements Runnable, Saver {
                 }
                 synchronized (this) {
                     try {
-                        active = false;
                         wait();
                     } catch (final InterruptedException e) {
                         log.warn("AsyncSaver got interrupted!");
                     }
                 }
             }
-            while (!active) {
-                try {
-                    con.getDatabase().getConnection();
-                    active = true;
-                } catch (final IllegalStateException illegalStateException) {
-                    log.warn("Failed to re-establish connection with the database! Trying again in %s second(s)..."
-                            .formatted(reconnectInterval / 1000), illegalStateException);
-                    try {
-                        Thread.sleep(reconnectInterval);
-                    } catch (final InterruptedException interruptedException) {
-                        log.warn("AsyncSaver got interrupted!", interruptedException);
-                        return;
-                    }
-                }
-            }
             final Record rec = queue.poll();
-            con.updateSQL(rec.type(), new Arguments(rec.args()));
+            if (rec != null) {
+                con.updateSQL(rec.type(), new Arguments(rec.args()));
+            }
         }
     }