+++ code/core/src/main/java/org/betonquest/betonquest/kernel/component/DatabaseComponent.java
@@ -1,16 +1,22 @@
 package org.betonquest.betonquest.kernel.component;
 
+import com.zaxxer.hikari.pool.HikariPool;
 import org.betonquest.betonquest.api.config.ConfigAccessor;
 import org.betonquest.betonquest.api.dependency.DependencyProvider;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.database.Connector;
 import org.betonquest.betonquest.database.Database;
+import org.betonquest.betonquest.database.DatabaseType;
 import org.betonquest.betonquest.database.MySQL;
 import org.betonquest.betonquest.database.SQLite;
 import org.betonquest.betonquest.lib.dependency.component.AbstractCoreComponent;
 import org.bukkit.plugin.Plugin;
+import org.jetbrains.annotations.Nullable;
 
+import java.io.IOException;
+import java.nio.file.Path;
+import java.sql.SQLException;
 import java.util.Set;
 
 /**
@@ -45,39 +51,15 @@ public class DatabaseComponent extends AbstractCoreComponent {
         final BetonQuestLoggerFactory loggerFactory = getDependency(BetonQuestLoggerFactory.class);
         final ConfigAccessor config = getDependency(ConfigAccessor.class);
         final Plugin plugin = getDependency(Plugin.class);
-
         final BetonQuestLogger log = loggerFactory.create(DatabaseComponent.class);
 
-        final boolean mySQLEnabled = config.getBoolean("mysql.enabled", true);
-        Database database = null;
-        if (mySQLEnabled) {
-            log.debug("Connecting to MySQL database");
-            final Database mySql = new MySQL(loggerFactory.create(MySQL.class, "Database"), plugin, config,
-                    config.getString("mysql.host"),
-                    config.getString("mysql.port"),
-                    config.getString("mysql.base"),
-                    config.getString("mysql.user"),
-                    config.getString("mysql.pass"));
-            try {
-                mySql.getConnection();
-                database = mySql;
-                this.mySql = true;
-                log.info("Successfully connected to MySQL database!");
-            } catch (final IllegalStateException e) {
-                log.warn("MySQL: " + e.getMessage(), e);
-            }
-        }
-        if (database == null) {
-            database = new SQLite(loggerFactory.create(SQLite.class, "Database"), plugin, config, "database.db");
-            if (mySQLEnabled) {
-                log.warn("No connection to the mySQL Database! Using SQLite for storing data as fallback!");
-            } else {
-                log.info("Using SQLite for storing data!");
-            }
-        }
+        final Path databasePropertiesPath = plugin.getDataFolder().toPath().resolve("database.properties");
+        final DatabaseConfig dbConfig = loadDatabaseConfig(config, log);
+
+        final Database database = resolveDatabase(dbConfig, loggerFactory, plugin, config, databasePropertiesPath, log);
 
         database.createTables();
-        final Connector connector = new Connector(config.getString("mysql.prefix"), database);
+        final Connector connector = new Connector(dbConfig.prefix, database);
 
         dependencyProvider.take(Connector.class, connector);
         dependencyProvider.take(DatabaseComponent.class, this);
@@ -90,5 +72,75 @@ public class DatabaseComponent extends AbstractCoreComponent {
      */
     public boolean usesMySQL() {
         return mySql;
+    }
+
+    private Database resolveDatabase(final DatabaseConfig dbConfig, final BetonQuestLoggerFactory loggerFactory,
+                                     final Plugin plugin, final ConfigAccessor config, final Path propertiesPath,
+                                     final BetonQuestLogger log) {
+        if (dbConfig.type == DatabaseType.MYSQL) {
+            log.debug("Connecting to MySQL database");
+            try {
+                final Database mySQL = new MySQL(loggerFactory.create(MySQL.class, "Database"), plugin, config, dbConfig, propertiesPath);
+                mySQL.getConnection();
+                this.mySql = true;
+                log.info("Successfully connected to MySQL database!");
+                return mySQL;
+            } catch (final HikariPool.PoolInitializationException | SQLException | IllegalStateException e) {
+                log.warn("Could not connect to MySQL database, falling back to SQLite: " + e.getMessage());
+                log.debug("MySQL connection error trace:", e);
+                log.warn("No connection to the MySQL Database! Using SQLite for storing data as fallback!");
+            }
+        }
+
+        log.debug("Connecting to SQLite database");
+        return new SQLite(loggerFactory.create(SQLite.class, "database"), plugin, config, "database.db", dbConfig);
+    }
+
+    private DatabaseConfig loadDatabaseConfig(final ConfigAccessor config, final BetonQuestLogger log) {
+        if (config.getBoolean("database.use_properties_file")) {
+            try {
+                return getDatabaseConfigWithNewConfig(config);
+            } catch (final IOException e) {
+                log.error("Failed to load 'database.properties' file! Using default.", e);
+            }
+        }
+        return getDatabaseConfigWithLegacyConfig(config);
+    }
+
+    private DatabaseConfig getDatabaseConfigWithNewConfig(final ConfigAccessor config) throws IOException {
+        final String prefix = config.getString("database.table_prefix", "betonquest_");
+        final DatabaseType type = DatabaseType.fromString(config.getString("database.type", "sqlite"));
+
+        return new DatabaseConfig(type, null, null, null, null, null, prefix);
+    }
+
+    private DatabaseConfig getDatabaseConfigWithLegacyConfig(final ConfigAccessor config) {
+        final int port = config.getInt("mysql.port", 3306);
+        final String host = config.getString("mysql.host", "");
+        final String base = config.getString("mysql.base", "");
+        final String user = config.getString("mysql.user", "");
+        final String password = config.getString("mysql.pass", "");
+
+        final String prefix = config.getString("database.table_prefix", "");
+        final DatabaseType type = DatabaseType.fromString(config.getString("database.type", "sqlite"));
+
+        return new DatabaseConfig(type, port, host, base, user, password, prefix);
+    }
+
+    /**
+     * Database configuration record.
+     *
+     * @param type   the database type
+     * @param port   the database port
+     * @param host   the database host address
+     * @param base   the database name
+     * @param user   the database username
+     * @param pass   the database password
+     * @param prefix the table prefix
+     */
+    public record DatabaseConfig(DatabaseType type, @Nullable Integer port, @Nullable String host,
+                                 @Nullable String base,
+                                 @Nullable String user, @Nullable String pass, String prefix) {
+
     }
 }